Overview of this update
Refactored app.py into Blueprints to improve modularity and maintainability, following the rubric’s guidelines.

Added CSRF token support to all form submissions for enhanced security.

Moved the login, signup, and password reset pages (and their assets) into a new login_system/ folder, and updated all internal paths accordingly.

Known issue

The login templates still need to be converted to use Jinja inheritance for consistent layout.

Next steps
Time is tight—please prioritize completing the remaining functional requirements so we can merge on schedule.
