+++ login_system/forgot_password.html
@@ -19,6 +19,12 @@
   </div>
 
   <script>
+    let csrfToken = '';
+    // Fetch CSRF token on page load
+    fetch('/api/csrf-token').then(res => res.json()).then(data => {
+      csrfToken = data.csrf_token;
+    });
+
     // Handle forgot-password form submission
     function submitForgotPassword(event) {
       event.preventDefault();
@@ -26,16 +32,17 @@
 
       fetch("/forgot-password", {
         method: "POST",
-        headers: { "Content-Type": "application/x-www-form-urlencoded" },
+        headers: { 
+          "Content-Type": "application/x-www-form-urlencoded",
+          'X-CSRFToken': csrfToken
+        },
         body: "email=" + encodeURIComponent(email)
       })
       .then(res => res.json())
       .then(data => {
         if (data.success) {
-          // Inform the user that reset link has been sent
           alert("Reset password email has been sent to your email.");
         } else {
-          // Show error if user not found or other issue
           alert("Error: " + (data.error || "User not found"));
         }
       })
